src/pyproject.toml
[tool.poetry]
name = "celery-django-example"
version = "0.1.0"
description = ""
authors = ["MATwave <stanislav.matveew@yandex.ru>"]
readme = "README.md"
packages = [{include = "celery_django_example"}]

[tool.poetry.dependencies]
python = "^3.10"
django = {extras = ["mail"], version = "^4.2.1"}
pre-commit = "^3.3.2"
install = "^1.3.5"
python-dotenv = "^1.0.0"
pydantic = "^1.10.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
